pyproject.toml
[tool.poetry]
name = "finsight-ai"
version = "0.1.0"
description = "Autonomous Financial Analyst Agent powered by RAG & LlamaIndex"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.10.15"
llama-index-llms-gemini = "^0.1.0"
llama-index-embeddings-huggingface = "^0.1.0"
llama-index-vector-stores-chroma = "^0.1.0"
google-generativeai = "^0.3.2"
streamlit = "^1.31.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
plotly = "^5.18.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
chromadb = "^0.4.22"
python-dotenv = "^1.0.0"
sec-edgar-downloader = "^5.0.0"
ragas = "^0.1.0"
pytest = "^8.0.0"
pytest-mock = "^3.12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

